Added sloppy function createMeta() combines the population data with the result from getTimeTrainMatrix() and cleans up the dataframe for classification use.
Corrected bias introduced by setting all INT_FLAG=0 patients to 180 days prior to threshold. Classification doesn't work as well as initially thought. Classification for INT_FLAG was found to work best without weighting on the day of the lab result and with the division between INT_FLAG=0 and INT_FLAG=1 set to > 0.515 goes to 1, otherwise 0.